fix(crawl): address CodeRabbit comments on PR #68 (LAB-2222)
Resolves review 4156714509:

- Rename browser test files to follow the engine_* convention:
  browser_test.go (integration) -> browser_integration_test.go, and
  browser_unit_test.go -> browser_test.go so the default-suite test
  file matches its source filename. CodeRabbit flagged the *_unit_
  suffix as non-standard.
- Add a nil-receiver guard to BrowserManager.SetCookies so calling
  the method on a nil *BrowserManager returns the same
  "browser not connected" error instead of panicking. Extend the
  default-suite test to cover this path
  (TestBrowserManager_SetCookies_NilReceiverReturnsError).
- In TestCrawler_CookieHeaderPropagatesAcrossRedirects, change the
  authed "/" handler to issue an actual 302 to /protected instead
  of returning an HTML link. This exercises the cookie-across-
  redirect path directly (the Spring Security scenario the test
  regresses), not just frontier link-following.
